+++ packages/router/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## UNRELEASED
 
+## 3.0.2
+
 ## 3.0.1
 
 ### Changed